requirements.txt
streamlit
langchain-openai
langchain-community
python-dotenv
ftfy
PyPDF2
pdfminer
cohere
faiss-cpu